非同期通信実装（いいね・コメント機能）

+++ app/controllers/favorites_controller.rb
@@ -6,14 +6,24 @@ class FavoritesController < ApplicationController
     favorite =
 current_user.favorites.new(book_id: book.id)
     favorite.save
+    respond_to do |format|
+      format.html { redirect_back fallback_location: root_path }
+      format.js
     redirect_back(fallback_location: books_path)
+    end
   end
 
   def destroy
-    book = Book.find(params[:book_id])
+    @book = Book.find(params[:book_id])
     favorite =
-current_user.favorites.find_by(book_id: book.id)
-    favorite.destroy
+current_user.favorites.find_by(book_id: @book.id)
+    if favorite.present?
+       favorite.destroy
+    end
+    respond_to do |format|
+      format.html { redirect_back fallback_location: books_path }
+      format.js
     redirect_back(fallback_location: books_path)
+    end
   end
 end
